cloud_scraper/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-logging==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
numpy==1.19.4
pandas==1.1.5
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
rsa==4.6
six==1.15.0
urllib3==1.26.2
